Updated to use texts from file.
Added a migrator that reads the texts from the file and stores them
in a level database. The texts are keyed on their index in the array
so it's important that we are careful when adding texts to the
database. This update also adds a `!text` command that will say a
random text from the db to the channel.